requirements.txt
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
click==7.1.2
codecov==2.1.3
coverage==5.1
Flask==1.1.2
Flask-API==2.0
idna==2.8
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.4.1
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
PyYAML==5.1.2
requests==2.22.0
requests-mock==1.8.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.6
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
